remove duplication in Devices::select() and Devices::enable()
