+++ msp430fr5739/delay.cpp
@@ -1,4 +1,4 @@
-#include <Energia.h>
+#include <Arduino.h>
 
 #include "device.h"
 #include "delay.h"
